# mes-root/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Mes Auto
  description: auto api
  contact:
    email: ixtf1984@gmail.com
  version: 1.0.0
security:
- bearerAuth: []
...
paths:
...
  /api/lineMachines/{id}/productPlan:
    get:
      operationId: productPlan
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineMachineProductPlan'
...
  /api/productPlanNotifies/{id}/exe:
    post:
      operationId: exe
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /api/productPlanNotifies/{id}/unExe:
    delete:
      operationId: unExe
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /api/productPlanNotifies/{id}/exeInfo:
    get:
      operationId: exeInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
  /api/productPlanNotifies/{id}/finish:
    put:
      operationId: finish
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
    delete:
      operationId: unFinish
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /api/productPlanNotifies/{id}/batchExe:
    post:
      operationId: get_16
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json: {}
  /api/productPlanNotifies/{id}:
    get:
      operationId: get_17
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPlanNotify'
    put:
      operationId: update_12
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPlanNotify'
  /api/productPlanNotifies:
    get:
      operationId: query_6
      parameters:
      - name: first
        in: query
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 50
      - name: q
        in: query
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: create_11
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPlanNotify'
  /api/productProcesses/{id}:
    get:
      operationId: get_18
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProcess'
    put:
      operationId: update_13
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProcess'
  /api/productProcesses:
    post:
      operationId: create_12
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductProcess'
  /api/products/{id}/productProcesses:
    get:
      operationId: productProcesses
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
  /api/products/{id}/dyeingInfo:
    get:
      operationId: dyeingInfo_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
    put:
      operationId: dyeingInfo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/products/{id}:
    get:
      operationId: get_19
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      operationId: update_14
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /api/products:
    get:
      operationId: query_7
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
    post:
      operationId: create_13
      requestBody:
        content:
          '*/*':
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
components:
  schemas:
...
    Batch:
      required:
      - batchNo
      - id
      - product
      - spec
      - tubeColor
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        product:
          $ref: '#/components/schemas/Product'
        batchNo:
          type: string
        silkWeight:
          minimum: 1
          type: number
          format: double
        centralValue:
          minimum: 1
          type: number
          format: double
        holeNum:
          minimum: 1
          type: integer
          format: int32
        spec:
          type: string
        tubeColor:
          type: string
        note:
          type: string
    Corporation:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
...
    Line:
      required:
      - id
      - name
      - workshop
      type: object
      properties:
        id:
          type: string
        workshop:
          $ref: '#/components/schemas/Workshop'
        name:
          type: string
        doffingType:
          type: string
          enum:
          - AUTO
          - MANUAL
          - DYEING_SAMPLE
          - PHYSICAL_INFO
    LineMachine:
      required:
      - id
      - line
      - spindleSeq
      type: object
      properties:
        id:
          type: string
        line:
          $ref: '#/components/schemas/Line'
        item:
          minimum: 1
          type: integer
          format: int32
        spindleNum:
          minimum: 1
          type: integer
          format: int32
        spindleSeq:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            type: integer
            format: int32
...
    Product:
      required:
      - code
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
    SapT001l:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        lgort:
          type: string
        lgobe:
          type: string
...
    SilkException:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    SilkNote:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mustFeedback:
          type: boolean
...
    Workshop:
      required:
      - code
      - corporation
      - id
      - name
      type: object
      properties:
        id:
          type: string
        corporation:
          $ref: '#/components/schemas/Corporation'
        name:
          type: string
        code:
          maxLength: 1
          minLength: 1
          type: string
        note:
          type: string
        sapT001ls:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsForeign:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
        sapT001lsPallet:
          type: array
          items:
            $ref: '#/components/schemas/SapT001l'
...
    FormConfig:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        formFieldConfigs:
          type: array
          items:
            $ref: '#/components/schemas/FormFieldConfig'
    FormFieldConfig:
      required:
      - inputType
      - name
      - valueType
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        valueType:
          type: string
          enum:
          - STRING
          - BOOLEAN
          - NUMBER
        inputType:
          type: string
          enum:
          - DEFAULT
          - SELECTION
        required:
          type: boolean
        multi:
          type: boolean
        selectOptions:
          type: array
          items:
            type: string
...
    LineMachineProductPlan:
      required:
      - batch
      - id
      type: object
      properties:
        id:
          type: string
        productPlanNotify:
          $ref: '#/components/schemas/ProductPlanNotify'
        lineMachine:
          $ref: '#/components/schemas/LineMachine'
        batch:
          $ref: '#/components/schemas/Batch'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    ProductPlanNotify:
      required:
      - batch
      - id
      - lineMachines
      - name
      - startDate
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum:
          - CHANGE_BATCH
          - SAMPLE
        name:
          type: string
        batch:
          $ref: '#/components/schemas/Batch'
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        lineMachines:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/LineMachine'
...
    ProductProcess:
      required:
      - exceptions
      - id
      - name
      - notes
      - product
      type: object
      properties:
        id:
          type: string
        product:
          $ref: '#/components/schemas/Product'
        name:
          type: string
        relateRoles:
          type: array
          items:
            type: string
            enum:
            - DOFFING
            - SUBMIT_DYEING_PREPARE
            - SUBMIT_DYEING_RESULT
            - SUBMIT_GRADE
            - SUBMIT_DYEING_SAMPLE
            - INSPECTION
            - PACKAGE_BOX
        sortBy:
          type: integer
          format: int32
        exceptions:
          maxItems: 2147483647
          minItems: 1
          type: array
          items:
            $ref: '#/components/schemas/SilkException'
        notes:
          type: array
          items:
            $ref: '#/components/schemas/SilkNote'
        formConfig:
          $ref: '#/components/schemas/FormConfig'
        mustProcess:
          type: boolean
...